# packages/core-schemas/openapi.yaml
openapi: 3.1.0
info:
  title: DreamRight API
  description: |
    AI-powered webtoon and short-form drama production API.

    ## Authentication
    All endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Workflow
    The typical workflow for creating a webtoon project is:
    1. Create a project (`POST /projects`)
    2. Expand a story prompt (`POST /projects/{id}/story`)
    3. Generate character assets (`POST /projects/{id}/characters/{id}/assets`)
    4. Generate location assets (`POST /projects/{id}/locations/{id}/assets`)
    5. Generate chapters sequentially (`POST /projects/{id}/chapters`)
    6. Generate panels for each chapter (`POST /projects/{id}/chapters/{id}/panels`)

    ## Asset Dependencies
    - Panel generation requires character portraits and location references to exist
    - Chapter N generation requires Chapter N-1 to exist (for story continuity)
    - Panel generation for Chapter N requires Chapter N-1 panels to exist (for visual continuity)

    ## Async Operations
    Long-running generation operations return `202 Accepted` with a job resource.
    Poll the job status endpoint or use webhooks for completion notifications.
  version: 1.0.0
  contact:
    name: DreamRight Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /projects/{projectId}/story:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Story]
      summary: Get story details
      operationId: getStory
      responses:
        '200':
          description: Story details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Story]
      summary: Create/expand story from prompt
      description: |
        Uses AI to expand a story prompt into a full story structure including:
        - Title, logline, and synopsis
        - Genre and tone
        - Story beats for each episode
        - Character definitions
        - Location definitions

        This is an async operation. Returns 202 with a job ID to poll for completion.
      operationId: createStory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoryRequest'
            example:
              prompt: "A shy high school girl discovers she can see ghosts"
              genre: "fantasy"
              episodes: 10
      responses:
        '202':
          description: Story expansion started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Story already exists (use PUT to replace)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Story]
      summary: Replace/re-expand story
      description: Replace existing story with a new expansion. Async operation.
      operationId: replaceStory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoryRequest'
      responses:
        '202':
          description: Story replacement started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

  # Character endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "unauthorized"
            message: "Authentication required"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "forbidden"
            message: "You do not have access to this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "rate_limited"
            message: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Genre:
      type: string
      enum: [romance, action, fantasy, thriller, slice_of_life, horror, comedy, drama, mystery, scifi]
    Tone:
      type: string
      enum: [comedic, dramatic, dark, lighthearted, romantic, suspenseful]
...
    JobStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    ErrorCode:
      type: string
      enum:
        - bad_request
        - validation_error
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - dependency_not_met
        - asset_exists
        - rate_limited
        - internal_error
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    JobResponse:
      type: object
      required: [job_id, status, poll_url]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        poll_url:
          type: string
          format: uri
          description: URL to poll for job status
...
    StoryBeat:
      type: object
      required: [beat, description]
      properties:
        beat:
          type: string
        description:
          type: string
    Story:
      type: object
      required: [id, title, genre, tone, episode_count]
      properties:
        id:
          type: string
        title:
          type: string
        logline:
          type: string
        genre:
          $ref: '#/components/schemas/Genre'
        tone:
          $ref: '#/components/schemas/Tone'
        themes:
          type: array
          items:
            type: string
        target_audience:
          type: string
        episode_count:
          type: integer
        synopsis:
          type: string
        story_beats:
          type: array
          items:
            $ref: '#/components/schemas/StoryBeat'
    CreateStoryRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          description: Story idea/prompt to expand
          minLength: 10
        genre:
          $ref: '#/components/schemas/Genre'
          description: Genre hint (optional, AI will decide if not specified)
        tone:
          $ref: '#/components/schemas/Tone'
          description: Tone hint (optional, AI will decide if not specified)
        episodes:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
          description: Target number of episodes

    # Character schemas
...